Virtual Invitation Creator/src/app/requirements.txt
flask
gunicorn
flask_sqlalchemy
pyjwt
bcrypt